src: sw_impl: added mbc3 write ram, works on hardware too but needs rework
While it also works flawlessly on the hardware I'm not happy with the way it is implemented now. There has to be some sort of flow-control to not overrun the RX buffer on the Zynq.
This will only be necessary for the standalone implementation so we need to #ifdef the relevant parts?