src: sw_impl: removed echoing, rewriting code for python/uart
Since the linuxapp is supposed to share most of the code with with the standalone app it is important to organize the code correctly.
However the python/uart frontend is not complete yet, so we will implement this completely first, and when it's time to work on the linuxapp the boundaries / common code will be much clearer so we can separate them later.